chore(lint): satisfy clippy 1.96 and port flac bench to symphonia 0.6
Rust 1.96 promoted several lints that -D warnings now rejects
(unnecessary_cast, collapsible_if, manual_div_ceil, manual_repeat_n,
needless_range_loop, len_zero); encoder internals and test helpers
with wide signatures get #[allow(clippy::too_many_arguments)].

The flac_rw bench still targeted the symphonia 0.5 API; the 0.6
release moved probing to Probe::probe, split decoders per track type
(make_audio_decoder + AudioDecoderOptions), and made next_packet
return Result<Option<Packet>>.

+++ src/flac/data.rs
@@ -314,7 +314,7 @@ mod tests {
     #[test]
     fn test_read_samples_multi_channel_shape() {
         let n = 8;
-        let channels: Vec<Vec<i32>> = (0..6).map(|ch| vec![(ch as i32) * 100; n]).collect();
+        let channels: Vec<Vec<i32>> = (0..6).map(|ch| vec![ch * 100; n]).collect();
         let audio = DecodedAudio::new(channels, 24, 96000);
 
         let sample_rate = NonZeroU32::new(96000).unwrap();
